refactor: simplify ClientProfileHeader and introduce ProfileRoleSelector component
- Removed unused state and menu handling from ClientProfileHeader, streamlining the component.
- Created ProfileRoleSelector component to manage user role selection, improving code organization and reusability.
- Updated UserProfileEntryPoint to utilize the new ProfileRoleSelector for role management.
