+++ modules/user-profile/components/tabs/user-contract/tabs/PersonalData/api/get-personal-data.ts
@@ -1,3 +1,4 @@
+import { UserRoleType } from "@/app/[locale]/(main)/client-profile/[id]/types";
 import { apiClient } from "@/config/axios-config";
 
 export type PersonalUserDataSectionT = {
@@ -11,6 +12,8 @@ export type PersonalUserDataSectionT = {
   country: string;
   country_id: number;
   nationalityRelation: string;
+  id: string;
+  user_types?: UserRoleType[];
 };
 
 type ResponseT = {
